leaderboard appears to be working
Each person now has a list of his/her bids per zip, city, and state.  For starters, I also compute the mean percent difference of those bids.

These objects can then be sorted and displayed (not done yet) to show, e.g. the highest mean percent difference within a certain zip code for people with more than N bids.  Yea!

This also gets recomputed whenever a house is marked as sold - the bid is added to the set of bids for that user for the geography of that house, and the overall metric is recomputed.  This updates the index with a minimal set of overhead and leaderboards are always up-to-date with the houses that are marked as sold or not.

The index can also be recomputed when the webapp starts in case new criteria are added, like a new geography or a new score type (like median vs. mean).